Legger database-helsesjekk i interfacet, i og med at den er lik i begge implementasjoner.
