+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/database/Database.kt
@@ -7,8 +7,14 @@ import no.nav.personbruker.dittnav.eventaggregator.common.exceptions.AggregatorB
 import no.nav.personbruker.dittnav.eventaggregator.common.exceptions.RetriableDatabaseException
 import no.nav.personbruker.dittnav.eventaggregator.common.exceptions.UnretriableDatabaseException
 import no.nav.personbruker.dittnav.eventaggregator.health.HealthCheck
+import no.nav.personbruker.dittnav.eventaggregator.health.HealthStatus
+import no.nav.personbruker.dittnav.eventaggregator.health.Status
 import org.postgresql.util.PSQLException
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import java.sql.*
+
+val log: Logger = LoggerFactory.getLogger(Database::class.java)
 
 interface Database: HealthCheck {
 
@@ -39,6 +45,18 @@ interface Database: HealthCheck {
                 operationToExecute()
             }
         }
+    }
+
+    override suspend fun status(): HealthStatus {
+        val serviceName = "Database"
+        return withContext(Dispatchers.IO) {
+            try {
+                dbQuery { prepareStatement("""SELECT 1""").execute() }
+                HealthStatus(serviceName, Status.OK, "200 OK")
+            } catch (e: Exception) {
+                log.error("Selftest mot databasen feilet", e)
+                HealthStatus(serviceName, Status.ERROR, "Feil mot DB")
+            }}
     }
 }
