+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/config/PostgresDatabase.kt
@@ -2,19 +2,12 @@ package no.nav.personbruker.dittnav.eventaggregator.config
 
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import no.nav.personbruker.dittnav.eventaggregator.common.database.Database
-import no.nav.personbruker.dittnav.eventaggregator.health.HealthStatus
-import no.nav.personbruker.dittnav.eventaggregator.health.Status
 import no.nav.vault.jdbc.hikaricp.HikariCPVaultUtil
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 
 class PostgresDatabase(env: Environment) : Database {
 
     private val envDataSource: HikariDataSource
-    private val log: Logger = LoggerFactory.getLogger(PostgresDatabase::class.java)
 
     init {
         envDataSource = createCorrectConnectionForEnvironment(env)
@@ -22,18 +15,6 @@ class PostgresDatabase(env: Environment) : Database {
 
     override val dataSource: HikariDataSource
         get() = envDataSource
-
-    override suspend fun status(): HealthStatus {
-        val serviceName = "Database"
-        return withContext(Dispatchers.IO) {
-            try {
-                dbQuery { prepareStatement("""SELECT 1""").execute() }
-                HealthStatus(serviceName, Status.OK, "200 OK")
-            } catch (e: Exception) {
-                log.error("Selftest mot databasen feilet", e)
-                HealthStatus(serviceName, Status.ERROR, "Feil mot DB")
-            }}
-    }
 
     private fun createCorrectConnectionForEnvironment(env: Environment): HikariDataSource {
         return when (ConfigUtil.isCurrentlyRunningOnNais()) {
@@ -84,5 +65,4 @@ class PostgresDatabase(env: Environment) : Database {
             return config
         }
     }
-
 }
